requirements.txt
netifaces==0.11.0
pyperclip==1.8.2
python_nmap==0.7.1
